migrate standardization function to seperate file
